Split promotion lifecycle integration tests
