+++ tests2/tests-map.json
@@ -2,13 +2,13 @@
   "$schema": "./tests-map.schema (informal): { generatedBy, censusTotal, buckets, methods, v2Native[{path,reason,execution}], journeys, entries[{file,bucket,method,replacement,rationale,migrated?,v2Path?,execution?}] }",
   "generatedBy": "scripts/testing-v2/gen-inventory.mjs, preserving reconciled v2Path/v2Native ownership",
   "note": "Re-run `node scripts/testing-v2/gen-inventory.mjs` after census changes. Classification overrides live in this generator; execution ownership is derived by test-map-execution.mjs.",
-  "censusTotal": 1142,
+  "censusTotal": 1143,
   "buckets": {
     "daily": 67,
     "v2-browser": 216,
     "v2-core": 512,
     "v2-dom": 144,
-    "v2-integration": 203
+    "v2-integration": 204
   },
   "methods": {
     "adapter": 276,
@@ -112,7 +112,16 @@
     },
     {
       "path": "tests2/integration/session-goal-promotion.test.ts",
-      "reason": "Integration coverage for promotion compensation, retries, lifecycle exclusion, exact checkout continuity, multi-repo cleanup, and sandbox ownership continuity.",
+      "reason": "Integration coverage for exact rollback, concurrent lifecycle exclusion, and exact checkout continuity.",
+      "execution": {
+        "runner": "vitest",
+        "tier": "unit",
+        "project": "integration"
+      }
+    },
+    {
+      "path": "tests2/integration/session-goal-promotion-recovery.test.ts",
+      "reason": "Integration coverage for recoverable reservations, retry finalization, multi-repo cleanup, and sandbox ownership continuity.",
       "execution": {
         "runner": "vitest",
         "tier": "unit",
